requirements.txt
pandas==1.0.5
PyMySQL==0.9.3
schedule==1.0.0
Werkzeug==0.16.0
tqdm==4.28.1
PureMySQL==0.0.10
Flask==1.1.2
requests==2.23.0
Flask_BasicAuth==0.2.0
APScheduler==3.7.0
SQLAlchemy==1.3.19
retrying==1.3.3
config==0.5.0.post0
paho_mqtt==1.5.1
python_dateutil==2.8.1
tinydb==4.4.0
